Changed Data type of the CurrentUser in Dal layer For space optimization

+++ BikeWaleOpr.UI/content/AddPricesPerFile.aspx.cs
@@ -118,7 +118,7 @@ namespace BikeWaleOpr.Content
                     cmd.Parameters.Add(DbFactory.GetDbParam("par_mumbaimetcorporaterto", DbType.Int64, Convert.DBNull));
                     cmd.Parameters.Add(DbFactory.GetDbParam("par_cityid", DbType.Int64, cityId));
                     cmd.Parameters.Add(DbFactory.GetDbParam("par_lastupdated", DbType.DateTime, DateTime.Now));
-                    cmd.Parameters.Add(DbFactory.GetDbParam("par_updatedby", DbType.Int64, -1));
+                    cmd.Parameters.Add(DbFactory.GetDbParam("par_updatedby", DbType.Int16, -1));
                     //run the command
                     MySqlDatabase.ExecuteNonQuery(cmd, ConnectionType.MasterDatabase);
                 }
